+++ src/HTTPie/Implement/RequestMapper.cs
@@ -4,7 +4,6 @@
 using HTTPie.Abstractions;
 using HTTPie.Models;
 using HTTPie.Utilities;
-using System.Net.Http.Headers;
 using System.Text;
 
 namespace HTTPie.Implement;
@@ -21,22 +20,12 @@ public sealed class RequestMapper : IRequestMapper
         if (!string.IsNullOrEmpty(requestModel.Body))
             request.Content = new StringContent(requestModel.Body, Encoding.UTF8,
                 httpContext.GetFlag(Constants.FlagNames.IsFormContentType)
-                    ? Constants.PlainTextMediaType
-                    : Constants.JsonMediaType);
+                    ? HttpHelper.TextPlainMediaType
+                    : HttpHelper.ApplicationJsonMediaType);
         if (requestModel.Headers is { Count: > 0 })
             foreach (var header in requestModel.Headers)
             {
-                if (Constants.ContentTypeHeaderName.EqualsIgnoreCase(header.Key))
-                {
-                    if (request.Content != null)
-                        request.Content.Headers.ContentType = MediaTypeHeaderValue.Parse(header.Value);
-                    continue;
-                }
-
-                if (HttpHelper.IsWellKnownContentHeader(header.Key) && request.Content != null)
-                    request.Content.Headers.TryAddWithoutValidation(header.Key, header.Value.ToArray());
-                else
-                    request.Headers.TryAddWithoutValidation(header.Key, header.Value.ToArray());
+                request.TryAddHeader(header.Key, header.Value.ToString());
             }
         return Task.FromResult(request);
     }