+++ src/HTTPie/Middleware/FollowRedirectMiddleware.cs
@@ -6,19 +6,14 @@ using HTTPie.Models;
 
 namespace HTTPie.Middleware;
 
-public sealed class FollowRedirectMiddleware : IHttpHandlerMiddleware
+public sealed class FollowRedirectMiddleware(HttpRequestModel requestModel) : IHttpHandlerMiddleware
 {
-    private readonly HttpRequestModel _requestModel;
+    private readonly HttpRequestModel _requestModel = requestModel;
 
     private static readonly Option<bool> FollowOption = new(new[] { "--follow", "-F" }, "The HTTP request should follow redirects");
     private static readonly Option<int> MaxRedirectsOption = new("--max-redirects", "Allowed max HTTP request redirect times");
 
-    public FollowRedirectMiddleware(HttpRequestModel requestModel)
-    {
-        _requestModel = requestModel;
-    }
-
-    public Task Invoke(HttpClientHandler httpClientHandler, Func<HttpClientHandler, Task> next)
+    public Task InvokeAsync(HttpClientHandler httpClientHandler, Func<HttpClientHandler, Task> next)
     {
         if (_requestModel.ParseResult.HasOption(FollowOption)
             || _requestModel.ParseResult.HasOption(DownloadMiddleware.DownloadOption))