+++ src/HTTPie/Middleware/ProxyMiddleware.cs
@@ -7,23 +7,18 @@ using System.Net;
 
 namespace HTTPie.Middleware;
 
-public sealed class ProxyMiddleware : IHttpHandlerMiddleware
+public sealed class ProxyMiddleware(HttpRequestModel requestModel) : IHttpHandlerMiddleware
 {
-    private readonly HttpRequestModel _requestModel;
+    private readonly HttpRequestModel _requestModel = requestModel;
     private static readonly Option<string> ProxyOption = new("--proxy", "Send request with proxy");
     private static readonly Option<bool> NoProxyOption = new("--no-proxy", "Disable proxy");
-
-    public ProxyMiddleware(HttpRequestModel requestModel)
-    {
-        _requestModel = requestModel;
-    }
 
     public Option[] SupportedOptions()
     {
         return new Option[] { ProxyOption, NoProxyOption };
     }
 
-    public Task Invoke(HttpClientHandler httpClientHandler, Func<HttpClientHandler, Task> next)
+    public Task InvokeAsync(HttpClientHandler httpClientHandler, Func<HttpClientHandler, Task> next)
     {
         if (_requestModel.ParseResult.HasOption(NoProxyOption))
         {