+++ src/HTTPie/Middleware/RequestHeadersMiddleware.cs
@@ -10,17 +10,17 @@ namespace HTTPie.Middleware;
 
 public sealed class RequestHeadersMiddleware : IRequestMiddleware
 {
-    public Task Invoke(HttpRequestModel requestModel, Func<HttpRequestModel, Task> next)
+    public Task InvokeAsync(HttpRequestModel requestModel, Func<HttpRequestModel, Task> next)
     {
         for (var i = requestModel.RequestItems.Count - 1; i >= 0; i--)
         {
             var item = requestModel.RequestItems[i];
             var index = item.IndexOf(':');
-            var key = item[..index];
             if (index > 0 && item.Length > (index + 1)
-                          && item[(index + 1)] != '='
-                          && key.IsMatch(Constants.ParamNameRegex))
+                          && item[(index + 1)] != '=')
             {
+                var key = item[..index];
+                if (!key.IsMatch(Constants.ParamNameRegex)) continue;
                 var value = item[(index + 1)..].Trim();
                 if (requestModel.Headers.TryGetValue(key, out var values))
                     requestModel.Headers[key] =