+++ src/HTTPie/Program.cs
@@ -2,42 +2,21 @@
 // Licensed under the MIT license.
 
 using HTTPie.Utilities;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using System.Text.Encodings.Web;
-using System.Text.Json;
+using System.Diagnostics;
+using WeihanLi.Common.Helpers.Hosting;
 
 var debugEnabled = args.Contains("--debug", StringComparer.OrdinalIgnoreCase);
-var logAsJson = args.Contains("--logAsJson", StringComparer.OrdinalIgnoreCase);
-var serviceCollection = new ServiceCollection();
-serviceCollection.AddLogging(builder =>
-{
-    if (logAsJson)
-    {
-        builder.AddJsonConsole(options =>
-        {
-            options.JsonWriterOptions = new JsonWriterOptions()
-            {
-                Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
-                Indented = true
-            };
-        });
-    }
-    else
-    {
-        builder.AddConsole();
-    }
-
-    builder.SetMinimumLevel(debugEnabled ? LogLevel.Debug : LogLevel.Warning);
-});
-serviceCollection.RegisterApplicationServices();
-await using var services = serviceCollection.BuildServiceProvider();
-
+var appBuilder = AppHost.CreateBuilder();
+appBuilder.Logging.AddDefaultDelegateLogger();
+appBuilder.Logging.SetMinimumLevel(debugEnabled ? LogLevel.Debug : LogLevel.Warning);
+appBuilder.Services.RegisterApplicationServices();
+var app = appBuilder.Build();
 // output helps when no argument or there's only "-h"/"/h"
 args = args switch
 {
-    [] => new[] { "--help" },
-    ["-h"] or ["/h"] => new[] { "--help" },
+    [] => ["--help"],
+    ["-h"] or ["/h"] => ["--help"],
     _ => args
 };
 if (args.Contains("--version"))
@@ -46,6 +25,12 @@ if (args.Contains("--version"))
     return 0;
 }
 
-var logger = services.GetRequiredService<ILogger>();
-logger.PrintInputParameters(args.StringJoin(";"));
-return await services.Handle(args);
+if (debugEnabled)
+{
+    app.Logger.PrintInputParameters(args.StringJoin(";"));
+#if DEBUG
+    if (!Debugger.IsAttached) Debugger.Launch();
+#endif
+}
+
+return await app.Services.Handle(args);