+++ tests/HTTPie.UnitTest/Middleware/RequestDataMiddlewareTest.cs
@@ -5,14 +5,9 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace HTTPie.UnitTest.Middleware;
 
-public class RequestDataMiddlewareTest
+public class RequestDataMiddlewareTest(IServiceProvider serviceProvider)
 {
-    private readonly IServiceProvider _serviceProvider;
-
-    public RequestDataMiddlewareTest(IServiceProvider serviceProvider)
-    {
-        _serviceProvider = serviceProvider;
-    }
+    private readonly IServiceProvider _serviceProvider = serviceProvider;
 
     [Theory]
     [InlineData("http://localhost:5000/api/values?name=test&hello=world")]
@@ -22,7 +17,7 @@ public class RequestDataMiddlewareTest
         await _serviceProvider.Handle(url, _ => Task.CompletedTask);
         var httpContext = _serviceProvider.GetRequiredService<HttpContext>();
         var middleware = new RequestDataMiddleware(httpContext);
-        await middleware.Invoke(httpContext.Request, _ => Task.CompletedTask);
+        await middleware.InvokeAsync(httpContext.Request, _ => Task.CompletedTask);
         Assert.Null(httpContext.Request.Body);
     }
 
@@ -34,7 +29,7 @@ public class RequestDataMiddlewareTest
         await _serviceProvider.Handle(input, _ => Task.CompletedTask);
         var httpContext = _serviceProvider.GetRequiredService<HttpContext>();
         var middleware = new RequestDataMiddleware(httpContext);
-        await middleware.Invoke(httpContext.Request, _ => Task.CompletedTask);
+        await middleware.InvokeAsync(httpContext.Request, _ => Task.CompletedTask);
         Assert.Null(httpContext.Request.Body);
     }
 
@@ -47,7 +42,7 @@ public class RequestDataMiddlewareTest
         await _serviceProvider.Handle(input, _ => Task.CompletedTask);
         var httpContext = _serviceProvider.GetRequiredService<HttpContext>();
         var middleware = new RequestDataMiddleware(httpContext);
-        await middleware.Invoke(httpContext.Request, _ => Task.CompletedTask);
+        await middleware.InvokeAsync(httpContext.Request, _ => Task.CompletedTask);
         Assert.Null(httpContext.Request.Body);
     }
 
@@ -60,7 +55,7 @@ public class RequestDataMiddlewareTest
         await _serviceProvider.Handle(input, _ => Task.CompletedTask);
         var httpContext = _serviceProvider.GetRequiredService<HttpContext>();
         var middleware = new RequestDataMiddleware(httpContext);
-        await middleware.Invoke(httpContext.Request, _ => Task.CompletedTask);
+        await middleware.InvokeAsync(httpContext.Request, _ => Task.CompletedTask);
         Assert.Null(httpContext.Request.Body);
     }
 
@@ -77,7 +72,7 @@ public class RequestDataMiddlewareTest
         await services.Handle(input, _ => Task.CompletedTask);
         var httpContext = services.GetRequiredService<HttpContext>();
         var middleware = new RequestDataMiddleware(httpContext);
-        await middleware.Invoke(httpContext.Request, _ => Task.CompletedTask);
+        await middleware.InvokeAsync(httpContext.Request, _ => Task.CompletedTask);
         Assert.NotNull(httpContext.Request.Body);
         Assert.Equal(@"{""Id"":1,""Name"":""Alice""}", httpContext.Request.Body);
     }